requirements.txt
streamlit
pandas
seaborn
plotly-express
numpy
scikit-learn==1.2.2
joblib
matplotlib
streamlit-option-menu
googletrans==4.0.0-rc1
altair
fuzzywuzzy
wordcloud
pycountry-convert
geopandas
tabulate
altair
mplcursors